LayoutTests/fast/html/script-must-not-run-when-child-is-removed-expected.txt

Adopting a script element must not run the script.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS script = scriptElement("html"); script.removeChild(textChild); didRunScript is false
PASS script = scriptElement("svg"); script.removeChild(textChild); didRunScript is false
PASS script = scriptElement("html"); script.removeChild(elementChild); didRunScript is false
PASS script = scriptElement("svg"); script.removeChild(elementChild); didRunScript is false
PASS script = scriptElement("html"); script.removeChild(commentChild); didRunScript is false
PASS script = scriptElement("svg"); script.removeChild(commentChild); didRunScript is false
PASS successfullyParsed is true

TEST COMPLETE
